refactor: 천 단위 구분 기호 추가
- MoneyTest: 구현 파일의 위치에 맞게 변경
