+++ src/test/java/domain/member/MoneyTest.java
@@ -1,6 +1,5 @@
-package domain.card;
+package domain.member;
 
-import domain.member.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -12,7 +11,7 @@ public class MoneyTest {
     @DisplayName("금액이 같으면 같은 객체로 취급한다 (VO)")
     @Test
     void equals_SameAmount_ReturnTrue() {
-        assertThat(new Money(10000)).isEqualTo(new Money(10000));
+        assertThat(new Money(10_000)).isEqualTo(new Money(10_000));
     }
 
     @DisplayName("베팅 금액이 0원 이하이면 예외가 발생한다")
@@ -25,11 +24,11 @@ public class MoneyTest {
     @DisplayName("수익률을 곱해 최종 수익금을 계산한다")
     @Test
     void multiply_EarningRate_ReturnProfit() {
-        Money bet = new Money(10000);
+        Money bet = new Money(10_000);
         double earningRate = 1.5;
 
         int profit = bet.calculateProfit(earningRate);
 
-        assertThat(profit).isEqualTo(15000);
+        assertThat(profit).isEqualTo(15_000);
     }
 }
